requirements.txt
pydantic==2.12.5
pillow==12.1.1
numpy==2.4.2
pytest==9.0.2
